Update completion binding
Change binding from readline "complete" command
to a macro, insert n at beginning of the line and then go back to
original position and activate completion.  This is a workaround if the
command already has some completion setting so that the binding should
always work.

Also, printing the table seems off again, needs more examination.
Lastly, if in vim mode, the mimic prompt is incorrect.